Renderer: Extracted renderer interface and backends to 'Renderer' directory
