+++ Libraries/Renderer/RendererBase/Precompiled.cpp
@@ -1,0 +1,2 @@
+// MIT Licensed (see LICENSE.md).
+#include "Precompiled.hpp"
