+++ Libraries/Renderer/RendererBase/Precompiled.hpp
@@ -1,0 +1,3 @@
+// MIT Licensed (see LICENSE.md).
+
+#include "RendererBaseStandard.hpp"
