+++ Libraries/Renderer/RendererBase/RendererBaseStandard.cpp
@@ -1,0 +1,7 @@
+// MIT Licensed (see LICENSE.md).
+#include "Precompiled.hpp"
+
+namespace Zero
+{
+
+}
