+++ Libraries/Renderer/RendererBase/RendererBaseStandard.hpp
@@ -1,0 +1,4 @@
+#include "Common/CommonStandard.hpp"
+
+#include "RendererEnumerations.hpp"
+#include "Renderer.hpp"
