+++ Libraries/Renderer/RendererEmpty/Precompiled.hpp
@@ -1,0 +1,3 @@
+// MIT Licensed (see LICENSE.md).
+
+#include "RendererEmptyStandard.hpp"
