+++ Libraries/Renderer/RendererEmpty/Renderer.cpp
