+++ Libraries/Renderer/RendererEmpty/RendererEmptyStandard.hpp
@@ -1,0 +1,3 @@
+#include "Renderer/RendererBase/RendererBaseStandard.hpp"
+
+//#include "Renderer.hpp"
