+++ Libraries/Renderer/RendererGL/RendererGLStandard.cpp
@@ -1,0 +1,7 @@
+// MIT Licensed (see LICENSE.md).
+#include "Precompiled.hpp"
+
+namespace Zero
+{
+
+}
