+++ Libraries/Renderer/RendererGL/Windows/OpenglRendererWindows.cpp
@@ -1,7 +1,7 @@
 // MIT Licensed (see LICENSE.md).
 
 #include "Precompiled.hpp"
-#include "../OpenGL/OpenglRenderer.hpp"
+#include "OpenglRenderer.hpp"
 
 namespace Zero
 {
