+++ Libraries/Support/SupportStandard.hpp
@@ -2,6 +2,7 @@
 #pragma once
 
 #include "Common/CommonStandard.hpp"
+#include "Renderer/RendererBase/RendererBaseStandard.hpp"
 #include "PlatformStandard.hpp"
 
 #include "Urls.hpp"
